pyproject.toml
[tool.poetry]
name = "api"
version = "0.5.6"
description = "Use Llama3 create an engaging quiz game incorporating TMDB API data"
authors = ["esx <king101125s@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
pydantic-settings = "^2.2.1"
google-cloud-aiplatform = ">=1.38"
jinja2 = "^3.1.3"
ruff = "^0.3.5"
pre-commit = "^3.7.0"
langchain-core = "^0.1.48"
langchain-community = "^0.0.36"
langchain-groq = "^0.1.3"
langchain-openai = "^0.1.6"
dashscope = "^1.19.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
